+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 ## Bug Fixes
 
 - PR #4386 Update Java package to 0.14
+- PR #4466 Fix merge key column sorting
 - PR #4402 Fix `cudf::strings::join_strings` logic with all-null strings and null narep
 - PR #4610 Fix validity bug in string scalar factory
 - PR #4570 Fixing loc ordering issue in dataframe